Add schematic sheet for the address decoder.
The address decoder is comprised to four GALs, two of which are mandatory
(they decode the RAM and ROM addresses plus the first 16 I/O device selects),
and two optional (they decode the an additional 16 I/O device selects,
respectively).  The input to each GAL is completely uniform, so any memory
map changes we decide to make can be done simply by reprogramming the GALs.