Add an interrupt controller.
The interrupt controller is made up of a PIA, plus 3 GALs (two as
interrupt priority encoders with individual interrupt enables,
and one as a multiplexer for those two encoders that provides a
read-only pending-interrupt register and pulls the system /IRQ line
low via a 2N7000).  The 2 I/O ports on the PIA are used for the
interrupt enables.