Oops, forgot to regen after this commit:
Simplify the fixed-vs-banked ROM selection logic.  Instead of using
2 '157 selectors, just use 2 '08 AND gate chips.  The bank signals
are AND'd with /FROMSEL to force them to 0 when the fixed ROM is
slected, and /FROMSEL and /BROMSEL are AND'd together to assert the
flash ROM chip select whenever either one of them is asserted.

This reverts the /FROMSEL and /BROMSEL logic in ADDRDEC0 back to the
original equations.