Re-work the interrupt controller to use discrete 74xx chips for mixing the enable signals and priority-encoding.
I made a major blunder in the INTENC GAL, and when I went to correct
it, it became clear that putting enable-mixing + priority encoding into
a single 22V10 wasn't going to fit.  At that point, the GAL was going
to be used either for priority encoding OR enable-mixing, and it just
didn't seem worth it.  So, INTENC goes away and is replaced by 74HC04s,
74HC00s, and 74HC148s.  The chip count kind of exploded here with 16
IRQs, but it fits on the board OK and the propagation delay for IRQ
signals doens't matter all that much.

Pay attention do how the 74HC148 is connected and look at the truth table
for that part carefully!